refactor: token metadata
