+++ packages/token-metadata/src/TokenMetaUtils.ts
@@ -5,20 +5,20 @@ import {
 import { getMappedTokensByName } from './tokens/mappings/mapByName'
 import { getMappedTokensByHash } from './tokens/mappings/mapByHash'
 import { getMappedTokensBySymbol } from './tokens/mappings/mapBySymbol'
-import { TokenMeta, TokenVerification, TokenType } from './types'
+import { TokenMetaBase, TokenVerification, TokenType } from './types'
 
 export class TokenMetaUtils {
-  protected tokens: Record<string, TokenMeta>
+  protected tokens: Record<string, TokenMetaBase>
 
-  protected tokensByErc20Address: Record<string, TokenMeta>
+  protected tokensByErc20Address: Record<string, TokenMetaBase>
 
-  protected tokensByCw20Address: Record<string, TokenMeta>
+  protected tokensByCw20Address: Record<string, TokenMetaBase>
 
-  protected tokensByHash: Record<string, TokenMeta>
+  protected tokensByHash: Record<string, TokenMetaBase>
 
-  protected tokensByName: Record<string, TokenMeta>
+  protected tokensByName: Record<string, TokenMetaBase>
 
-  constructor(tokens: Record<string, TokenMeta>) {
+  constructor(tokens: Record<string, TokenMetaBase>) {
     this.tokens = getMappedTokensBySymbol(tokens)
     this.tokensByErc20Address = getMappedTokensByErc20Address(this.tokens)
     this.tokensByCw20Address = getMappedTokensByCw20Address(this.tokens)
@@ -32,7 +32,7 @@ export class TokenMetaUtils {
    * - BaseDenom based on the ibc hash
    * - Variation of a symbol for multiple versions of the same token (ex: USDC, USDCet, USDCso)
    */
-  getMetaBySymbol(symbol: string): TokenMeta | undefined {
+  getMetaBySymbol(symbol: string): TokenMetaBase | undefined {
     const { tokens: tokensBySymbol } = this
     const tokenSymbol = symbol.toUpperCase() as keyof typeof tokensBySymbol
 
@@ -48,7 +48,7 @@ export class TokenMetaUtils {
     }
   }
 
-  getMetaByFactory(denom: string): TokenMeta | undefined {
+  getMetaByFactory(denom: string): TokenMetaBase | undefined {
     const [symbolOrName, creatorAddress] = denom.split('/').reverse()
     const tokenMeta =
       this.getMetaByName(symbolOrName) || this.getMetaBySymbol(symbolOrName)
@@ -57,21 +57,15 @@ export class TokenMetaUtils {
       return
     }
 
-    if (tokenMeta.tokenFactories) {
-      const tokenFactory = tokenMeta.tokenFactories.find(
-        (tokenFactory) => tokenFactory.creator === creatorAddress,
-      )
-
-      if (tokenFactory) {
-        return {
-          ...tokenMeta,
-          tokenType: TokenType.TokenFactory,
-          tokenVerification: TokenVerification.Verified,
-        }
-      }
+    if (!tokenMeta.tokenFactories) {
+      return
     }
 
-    if (tokenMeta.tokenFactory?.creator !== creatorAddress) {
+    const tokenFactory = tokenMeta.tokenFactories.find(
+      (tokenFactory) => tokenFactory.creator === creatorAddress,
+    )
+
+    if (!tokenFactory) {
       return
     }
 
@@ -82,13 +76,13 @@ export class TokenMetaUtils {
     }
   }
 
-  getMetaByAddress(address: string): TokenMeta | undefined {
+  getMetaByAddress(address: string): TokenMetaBase | undefined {
     return address.startsWith('0x')
       ? this.getMetaByErc20Address(address)
       : this.getMetaByCw20Address(address)
   }
 
-  getMetaByCw20Address(address: string): TokenMeta | undefined {
+  getMetaByCw20Address(address: string): TokenMetaBase | undefined {
     const { tokensByCw20Address } = this
     const contractAddress =
       address.toLowerCase() as keyof typeof tokensByCw20Address
@@ -112,7 +106,7 @@ export class TokenMetaUtils {
       : undefined
   }
 
-  getMetaByErc20Address(address: string): TokenMeta | undefined {
+  getMetaByErc20Address(address: string): TokenMetaBase | undefined {
     const { tokensByErc20Address } = this
     const contractAddress =
       address.toLowerCase() as keyof typeof tokensByErc20Address
@@ -170,7 +164,7 @@ export class TokenMetaUtils {
     }
   }
 
-  getMetaByHash(hash: string): TokenMeta | undefined {
+  getMetaByHash(hash: string): TokenMetaBase | undefined {
     const { tokensByHash } = this
     const ibcHash = hash
       .toUpperCase()
@@ -191,7 +185,7 @@ export class TokenMetaUtils {
       : undefined
   }
 
-  getMetaByName(name: string): TokenMeta | undefined {
+  getMetaByName(name: string): TokenMetaBase | undefined {
     const { tokensByName } = this
     const tokenName = name.toLowerCase() as keyof typeof tokensByName
