+++ packages/token-metadata/src/index.ts
@@ -5,7 +5,6 @@ import { TokenMetaUtilsFactory } from './TokenMetaUtilsFactory'
 export * from './ibc'
 export * from './types'
 export * from './utils'
-export * from './tokens/canonical'
 
 export const tokenMetaUtils = TokenMetaUtilsFactory.make()
 export const tokenFactory = TokenFactory.make()
