+++ packages/token-metadata/src/tokens/mappings/mapByName.ts
@@ -1,10 +1,10 @@
-import { TokenMeta } from '../../types'
+import { TokenMetaBase } from '../../types'
 
-export const getMappedTokensByName = (tokens: Record<string, TokenMeta>) =>
+export const getMappedTokensByName = (tokens: Record<string, TokenMetaBase>) =>
   (Object.keys(tokens) as Array<keyof typeof tokens>).reduce(
     (result, token) => ({
       ...result,
       [tokens[token].name!.toLowerCase()]: tokens[token],
     }),
     {},
-  ) as Record<string, TokenMeta>
+  ) as Record<string, TokenMetaBase>
