+++ packages/token-metadata/src/tokens/tokens/cw20-tokens.ts
@@ -1,14 +1,11 @@
-import { TokenMeta, TokenType } from '../../types'
+import { TokenMetaBase, TokenType } from '../../types'
 
 /** for testnet purposes only */
 export default {
   'INJ-MASTER': {
     name: 'INJ Master',
     logo: 'mito-inj-master-lp.png',
-    symbol: 'INJ-MASTER-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -21,10 +18,7 @@ export default {
   'INJ-BOOST': {
     name: 'INJ Boost',
     logo: 'mito-inj-boost-lp.png',
-    symbol: 'INJ-BOOST-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -37,10 +31,7 @@ export default {
   'XAU-BOOST': {
     name: 'XAU Boost',
     logo: 'mito-xau-boost-lp.png',
-    symbol: 'XAU-BOOST-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -53,10 +44,7 @@ export default {
   'WETH-MASTER': {
     name: 'WETH Master',
     logo: 'mito-weth-master-lp.png',
-    symbol: 'WETH-MASTER-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -69,10 +57,7 @@ export default {
   'ETH-BOOST': {
     name: 'ETH Boost',
     logo: 'mito-eth-boost-lp.png',
-    symbol: 'ETH-boost-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -85,10 +70,7 @@ export default {
   'ATOM-MASTER': {
     name: 'ATOM Master',
     logo: 'mito-atom-master-lp.png',
-    symbol: 'ATOM-MASTER-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -101,10 +83,7 @@ export default {
   'ATOM-BOOST': {
     name: 'ATOM Boost',
     logo: 'mito-atom-boost-lp.png',
-    symbol: 'ATOM-BOOST-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -117,10 +96,7 @@ export default {
   'USDC-MASTER': {
     name: 'USDC Master',
     logo: 'mito-usdc-master-lp.png',
-    symbol: 'USDC-MASTER-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -133,10 +109,7 @@ export default {
   'PROJ-MASTER': {
     name: 'PROJ Master',
     logo: 'mito-proj-master-lp.png',
-    symbol: 'PROJ-MASTER-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -149,10 +122,7 @@ export default {
   'GBP-MASTER': {
     name: 'GBP Master',
     logo: 'mito-gbp-master-lp.png',
-    symbol: 'GBP-MASTER-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -165,10 +135,7 @@ export default {
   'EUR-MASTER': {
     name: 'EUR Master',
     logo: 'mito-eur-master-lp.png',
-    symbol: 'EUR-MASTER-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -181,10 +148,7 @@ export default {
   'STINJ-BOOST': {
     name: 'stINJ Boost',
     logo: 'mito-stInj-boost-lp.png',
-    symbol: 'stINJ-BOOST-LP',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
       decimals: 18,
@@ -197,12 +161,10 @@ export default {
   ZEN: {
     name: 'ZEN',
     logo: 'zen.svg',
-    symbol: 'ZEN',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
+      symbol: 'ZEN',
       decimals: 18,
       address: 'uzen',
       tokenType: TokenType.Cw20,
@@ -212,12 +174,10 @@ export default {
   PROJ: {
     name: 'PROJ',
     logo: 'proj.png',
-    symbol: 'PROJ',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
+      symbol: 'PROJ',
       decimals: 18,
       address: 'proj',
       tokenType: TokenType.Cw20,
@@ -227,12 +187,10 @@ export default {
   PROJX: {
     name: 'PROJX',
     logo: 'projx.png',
-    symbol: 'PROJX',
-    decimals: 18,
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
+      symbol: 'PROJX',
       decimals: 18,
       address: 'projx',
       tokenType: TokenType.Cw20,
@@ -241,16 +199,14 @@ export default {
 
   DEMO: {
     name: 'Demo Coin',
-    symbol: 'DEMO',
-    decimals: 6,
     logo: 'injective-v3.png',
     coinGeckoId: '',
-    tokenType: TokenType.Cw20,
 
     cw20: {
+      symbol: 'DEMO',
       decimals: 18,
       address: 'demo',
       tokenType: TokenType.Cw20,
     },
   },
-} as Record<string, TokenMeta>
+} as Record<string, TokenMetaBase>
