+++ packages/token-metadata/src/utils/canonical.ts
