+++ packages/token-metadata/test/unit/tokens.spec.ts
@@ -13,11 +13,11 @@ describe('TokenMetadata', () => {
     expect(tokenMeta.ibc).toEqual({
       decimals: 6,
       isNative: true,
+      symbol: 'LUNA',
       baseDenom: 'uluna',
       path: 'transfer/channel-4',
       channelId: 'channel-4',
       hash: 'B8AF5D92165F35AB31F3FC7C7B444B9D240760FA5D406C49D24862BD0284E395',
-      tokenType: 'ibc',
     })
   })
 
@@ -81,16 +81,14 @@ describe('TokenMetadata', () => {
     const USDCPeggyDenom = 'peggy0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48'
 
     test('fetches correct value for peggyUsdc', () => {
-      const tokenMeta = tokenFactory.toToken(USDCetFactoryDenom) as Token
+      const tokenMeta = tokenFactory.toToken(USDCPeggyDenom) as Token
 
       expect(tokenMeta).toEqual(
         expect.objectContaining({
-          cw20: {
+          erc20: {
             decimals: 6,
-            symbol: 'USDCet',
-            source: 'ethereum-wormhole',
-            address: 'inj1q6zlut7gtkzknkk773jecujwsdkgq882akqksk',
-            tokenType: 'cw20',
+            symbol: 'USDC',
+            address: '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48',
           },
         }),
       )
@@ -106,7 +104,6 @@ describe('TokenMetadata', () => {
             symbol: 'USDCet',
             source: 'ethereum-wormhole',
             address: 'inj1q6zlut7gtkzknkk773jecujwsdkgq882akqksk',
-            tokenType: 'cw20',
           },
         }),
       )
@@ -122,7 +119,6 @@ describe('TokenMetadata', () => {
             symbol: 'USDCso',
             source: 'solana',
             address: 'inj12pwnhtv7yat2s30xuf4gdk9qm85v4j3e60dgvu',
-            tokenType: 'cw20',
           },
         }),
       )
@@ -137,7 +133,6 @@ describe('TokenMetadata', () => {
             address: '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48',
             decimals: 6,
             symbol: 'USDC',
-            tokenType: 'erc20',
           },
         }),
       )
